minor template update

+++ algo/template.cpp
@@ -62,10 +62,10 @@ typedef unsigned long long int ull;
 #define rep(...) GET_MACRO(__VA_ARGS__, forsn, qwe, forn)(__VA_ARGS__)
 #define repll(...) GET_MACRO(__VA_ARGS__, forsnll, qwe, fornll)(__VA_ARGS__)
 #define qwe(r,t,y)
-#define forn(i, n) for (int i = 0; i < n; i++)
-#define fornll(i, n) for (ll i = 0ll; i < n; i++)
-#define forsn(i, st, end, d) for(int i = st; (d>0?i<=end:i>=end); i+=d)
-#define forsnll(i, st, end, d) for(ll i = st; (d>0?i<=end:i>=end); i+=(ll)d)
+#define forn(i, n) for (int i = 0; i < (int)n; i++)
+#define fornll(i, n) for (ll i = 0ll; i < (ll)n; i++)
+#define forsn(i, st, end, d) for(int i = st; (d>0?i<=(int)end:i>=(int)end); i+=d)
+#define forsnll(i, st, end, d) for(ll i = st; (d>0?i<=(ll)end:i>=(ll)end); i+=(ll)d)
 #define all(x) (x).begin(), (x).end()
 #define pass (void)0
 #define space " "
